+++ src/super_image/models/__init__.py
@@ -10,3 +10,9 @@ from .masa.configuration_masa import MasaConfig
 from .masa.modeling_masa import MasaModel
 from .carn.configuration_carn import CarnConfig
 from .carn.modeling_carn import CarnModel
+from .jiif.configuration_jiif import JiifConfig
+from .jiif.modeling_jiif import JiifModel
+from .liif.configuration_liif import LiifConfig
+from .liif.modeling_liif import LiifModel
+from .smsr.configuration_smsr import SmsrConfig
+from .smsr.modeling_smsr import SmsrModel